tests: move Git-related routines to support
